requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
datasets==3.5.0
dill==0.3.8
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
Flask-Dance==7.1.0
frozenlist==1.5.0
fsspec==2024.12.0
gradio==5.25.0
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
multiprocess==0.70.16
numpy==2.2.4
oauthlib==3.2.2
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pfzy==0.3.4
pillow==11.2.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
ruff==0.11.5
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tokenizers==0.21.1
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.51.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
URLObject==2.4.3
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.19.0
